--- src/components/Test.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Test = () => {
-  const a = 1;
-  console.log('hello');
-
-  return (
-    <div>
-      <div>hi</div>
-    </div>
-  );
-};
-
-export default Test;
