+++ src/components/common/Slick.js
