+++ src/components/map/MapSlick.js
