--- src/pages/SearchTypeList.js
@@ -1,66 +1,0 @@
-/* eslint-disable react/jsx-props-no-spreading */
-import React from 'react';
-import { useSelector } from 'react-redux';
-import { Container, Grid } from '../elements';
-import SearchBar from '../components/SearchBar';
-import SelectedCategory from '../components/SelectedCategory';
-import { Slick } from '../components/Slick';
-import ListCard from '../components/ListCard';
-import ContentsTitle from '../components/ContentsTitle';
-import Header from '../components/Header';
-import Navbar from '../components/Navbar';
-
-const SearchTypeList = () => {
-  const conditionPlaces = useSelector(state => state.place.conditionPlaces);
-  const inSideList = conditionPlaces && conditionPlaces.insidePlaces;
-  const outSideList = conditionPlaces && conditionPlaces.outSidePlaces;
-  const srcList = [
-    { src: 'https://t1.daumcdn.net/cfile/tistory/213C9A345225669622' },
-    { src: 'https://img.siksinhot.com/article/1613970568705496.jpg' },
-    {
-      src: 'https://mblogthumb-phinf.pstatic.net/MjAxOTA4MjZfMjk5/MDAxNTY2Nzg2MzQzMTE2.Gf-bdbf4G8JpcXijSij4ydO7dKhCpeTUCISRwXDMI90g.Yeoe2X4MKjwyFStB4Vqm2FkTmlLEnx77Sihzs97-reog.JPEG.flourish12/output_2154676882.jpg?type=w800',
-    },
-    { src: 'https://cdn.st-news.co.kr/news/photo/202105/2097_3508_4043.jpg' },
-    {
-      src: 'https://www.artinsight.co.kr/data/tmp/2102/20210224004559_fnyxevid.jpg',
-    },
-    {
-      src: 'https://img1.daumcdn.net/thumb/R720x0/?fname=http%3A%2F%2Ft1.daumcdn.net%2Fliveboard%2Fdailylife%2Fd069f8ac17de4c7c91c0dc056d452779.jpg',
-    },
-    {
-      src: 'https://t1.daumcdn.net/thumb/R720x0/?fname=http://t1.daumcdn.net/brunch/service/user/2UUt/image/Prcj7-b1i0zgV78lpTYYBdP4GRw.jpg',
-    },
-    {
-      src: 'https://img.wkorea.com/w/2021/02/style_602cfee44c823-970x1200.jpg',
-    },
-  ];
-  return (
-    <>
-      <Container>
-        <Header _type="search" _back _content="검색 결과" _map _search />
-        {/* <SelectedCategory tag={selected[0]} /> */}
-        <ContentsTitle title="실내" />
-        <Grid>
-          <Slick>
-            {inSideList &&
-              inSideList.map(info => {
-                return <ListCard info={info} />;
-              })}
-          </Slick>
-        </Grid>
-        <Grid padding="0 0 74px 0">
-          <ContentsTitle title="실외에서 시원한 바람과 함께" />
-          <Slick>
-            {outSideList &&
-              outSideList.map(info => {
-                return <ListCard info={info} />;
-              })}
-          </Slick>
-        </Grid>
-      </Container>
-      <Navbar />
-    </>
-  );
-};
-
-export default SearchTypeList;
