+++ src/redux/configureStore.js
@@ -6,18 +6,14 @@ import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit';
 import { logger } from 'redux-logger';
 
 /* slice module */
-import counterSlice from './modules/counterSlice';
 import placeSlice from './modules/placeSlice';
-import selectedSlice from './modules/selected';
 import loadedSlice from './modules/loadedSlice';
 
 export const history = createBrowserHistory();
 
 /* Reducer combine */
 const reducer = combineReducers({
-  counter: counterSlice.reducer,
   place: placeSlice.reducer,
-  selected: selectedSlice.reducer,
   loaded: loadedSlice.reducer,
   router: connectRouter(history),
 });
