--- src/shared/api/detailApi.js
@@ -1,7 +1,0 @@
-import api from './index';
-
-const getDetail = params => {
-  return api.get(`/posts/${params}`);
-};
-
-export { getDetail };
