+++ src/shared/api/placeApi.js
@@ -1,14 +1,13 @@
 import api from './index';
 
-const getMainList = () => {
-  return api.get('main');
-};
+const getMainList = () => api.get('main');
 
 const getSearchCondition = params => {
-  console.log('params', params);
   return api.get(
     `search/condition?weather=${params.weather}&category=${params.category}&num=${params.num}&gender=${params.gender}`,
   );
 };
 
-export { getMainList, getSearchCondition };
+const getPlaceDetail = params => api.get(`/posts/${params}`);
+
+export { getMainList, getSearchCondition, getPlaceDetail };
