--- src/shared/api/roomApi.js
@@ -1,7 +1,0 @@
-import api from './index';
-
-const getRoomList = params => {
-  return api.get(`/roomList`);
-};
-
-export { getRoomList };
